backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
simple-salesforce
pyrfc
websockets
